+++ Source/Backend/mekongocop-server/src/main/java/com/mekongocop/mekongocopserver/entity/OrderItem.java
@@ -2,10 +2,14 @@ package com.mekongocop.mekongocopserver.entity;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "order_items")
 public class OrderItem {
-    public OrderItem(int order_item_id, int quantity, int price, Order order, Product product) {
+
+
+    public OrderItem(int order_item_id, int quantity, BigDecimal price, Order order, Product product) {
         this.order_item_id = order_item_id;
         this.quantity = quantity;
         this.price = price;
@@ -29,11 +33,11 @@ public class OrderItem {
         this.quantity = quantity;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 
@@ -57,7 +61,7 @@ public class OrderItem {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int order_item_id;
     private int quantity;
-    private int price;
+    private BigDecimal price;
 
     @ManyToOne
     @JoinColumn(name = "order_id")
